fix(ionic-react): update schematics to better handle adding Nx plugins
Due to recent changes, a user would encounter a "collection not found" error if they did not manually add `@nrwl/react` before using the `@nxtend/ionic-react` plugin.

This change now adds a requirement for users to execute the `init` schematic before the `application` schematic, but seems to be the best compromise at the moment.
