+++ packages/docs/docs/ionic-react/schematics/init.md
@@ -13,6 +13,8 @@ This schematic adds the following dependencies:
 ```
 @ionic/react
 @ionic/react-router
+@nrwl/react
+@nxtend/capacitor
 @testing-library/user-event
 @testing-library/jest-dom
 @testing-library/cypress
