+++ packages/ionic-react/collection.json
@@ -2,7 +2,6 @@
   "$schema": "../../node_modules/@angular-devkit/schematics/collection-schema.json",
   "name": "IonicReact",
   "version": "0.0.1",
-  "extends": ["@nrwl/react"],
   "schematics": {
     "init": {
       "factory": "./src/schematics/init/schematic",
