+++ packages/ionic-react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nxtend/ionic-react",
-  "version": "3.0.3",
+  "version": "3.0.4-dev",
   "description": "An Nx plugin for developing Ionic React applications and libraries",
   "author": {
     "name": "Devin Shoemaker",
@@ -18,7 +18,5 @@
   "ng-update": {
     "migrations": "./migrations.json"
   },
-  "dependencies": {
-    "@nxtend/capacitor": "*"
-  }
+  "dependencies": {}
 }
