--- packages/ionic-react/src/schematics/init/schema.d.ts
@@ -1,3 +1,0 @@
-export interface InitSchematicSchema {
-  capacitor: boolean;
-}
